requirements.txt
streamlit
seaborn
numpy
pandas
plotly-express
matplotlib